+++ src/main/java/com/project/email/scheduler/audit/audit_log/dao/AuditLogRepo.java
@@ -2,7 +2,7 @@ package com.project.email.scheduler.audit.audit_log.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditLogRepo extends JpaRepository<AuditLog, Long>,
+public interface AuditLogRepo extends JpaRepository<AuditLog, Integer>,
         AuditLogCustomRepo<AuditLog> {
 
 }
